ATR-803: refactoring - moved out collection utils dedicated to remove boxing for struct collections
